Ensure vector fits within limits in sample
